Improve interaction drawing as per issue #6
This improves the drawing of interactions a great deal, however they are not
perfect. In some cases it's a bit unclear as to where the interaction is going.
I will have to improve it in the future but for now this is good. The method is
as follows.

I'm going to detail my thoughts so I have a chance of figuring out why I did
this a couple months.

We draw a line between the center of the letters. For some reason using height/4
works better to get the vertical center than just height/2. It produces a more
pleasing location.

We then draw a right triangle using the direct line as the hypotenuse. We then
draw a circle of radius ntRadius from the first point. The line that is parallel
the y line that goes through the intersection of the direct line and circle
splits the sides of the triangle. We then use this splitting and a^2+b^2=c^2 to
determine the x and y coordinates of the point to start drawing at. Basically if
we label the direct line d, the radius r and the horizontal distance x we can
find the horizontal distance a by:

d/r = x-a/a

so a is

a = x*r / (d+r);

We then get b by

r^2 = a^2 + b^2

That would be a lot clearer with pictures.