Reorganize the library.
This commit reorganises the structure of this library. Things are now split into
separate functions in separate files which are concated together as needed. This
is hopefully a lot more useful than before. Also things can now be configured
much easier, I think. It doesn't completely work yet but this is the structure
I'm going with for now so I'll commit it.