Fix scales and selecting with the brush.
The brush had been using the x property of the objects. This is not good since
the x attribute is for the airport diagram. Instead we assign an __x property
to indicate the location. We don't override the x property otherwise we won't be
able to switch back to airport view.

Also we use an identity scale to make life easy.