Fix bug in drawing boxes.
Sometimes we are drawing motifs but we don't draw all because we stop too
early. Basically, if one nt was missing in we would stop drawing the motifs
because we didn't have a bounding for that motif. This would then prevent all
future motifs from being boxed. We now log this problem and give it a default
bounding box. This approach can be cleaned up.