Reorganize plugin
Anton suggests that the plugin should have a single entry point to set
everything up. I changed things around to match this. Now the .rna2d function
takes an init object which will init each of the parts in the plugin. In
addition, each part is now a property of the rna2d function. Not sure if
exposing them is a good idea, but this does make for simple organization.