Plot is a function of no arguments and fix drawing
Plot does not need to take the selection as an argument that only makes things
more complicated. In addition, drawing was not happening because we did not
select the top level stuff correctly. This fixes it by appending the svg as
needed as well as setting the width and the like.