Close #22.
We now always draw the arcs correctly. The trick to it is to check if we are
drawing across the circle. If we are then we need to deal with it, here I've
chosen to keep sweep 0 and change the order of the endpoints. It would also be
possible to change sweep instead.