Fixes #27.
This merges the functions of the example-page into the development view. This
should make for better development as the page I look at now is more like the
functions that are actually used. In addition it gives me ideas about what
other functions would be helpful in the jquery plugin.