Fixes #28
This was caused by messing with the width and height each time I was drawing. I
fixed that and moved the defining teh domain and range into the main section.
This is useful I think. In addition we don't have to worry about ever needing
to mess with the size now.