Fixes #30.
Motifs can now be toggled by type like with interactions. I've created a
togglable part of the library which is applied to nucleotides, motifs and
interactions which can toggle the display of things based upon their class. It
adds some common functions, like all, show, hide and such. The fact that it uses
a class is moderately important right now. Changing that is possible but will
pose problems.

This is also a step in the direction of dealing with #29. Right now it does not
because we clear out all the old information about what is stored each time we
redraw. That will be fixed though.