Minor cleanups.
The way everything was drawn was a bit odd. We would call a function on the
selection to do everything. It's easier to just do what we want to the
selection directly without the need for some callback. This reorganizes things
like that. Also since the code for the two scales is the same we create a
function for that.