Refactor main drawing function
We add a redraw function which redraws. This will be needed for zooming, maybe
later I can do some prefilter, if needed, to remove extra elements. In addition,
we refactor what is done to draw elements. For example, we do not need to add
the data like sequence and nts to nucleotides, interactions and such. We should do
filtering not using css selectors, but instead using data attributes. This
change forces that. We change the standardBuild function to make it simpiler to
build things because of these changes.