Fixes #3
We now have a basic zoom working. This allows us to focus in on one region to
see it clearly. I still should add more features like labels that only appear
on high zoom, or letters in circular that appear at high zoom and the like. But
this is a good enough start.