Use bounding box instead of indexOf to check nts
We need to check if the nucleotides in an interaction actually exist. To do
this we now get the bounding box instead of getting the index. We no longer
compute the indexes of each nucleotide so this is a better idea. In addition,
if I ever do optimize to not draw things this would be a better check.