Update airport to new system
This changes teh airport view to use the new setup. This includes changes such
as drawing chains, though the airport doesn't care too much about this, and the
new way to generate domains. In addition, I don't rely on the this object so
much in highlight functions. We are passed the data so there is no need to get
the data using this.