Update circular drawing to new chains setup
Since we now have multiple chains we need to change how circular diagrams are
drawn. They have to have gaps between chains. This updates to make that
possible. In addition, I have tried to simplify this view. The code is still
overly complicated and hard to follow, but I think this is a little better.