Generate circular diagrams with chain breaks
We now generate circular with chain breaks correctly. This is also a
reorganization of the circular view as it is messy. It still is messy but at
least it's better mostly. In addition, we add a preprocesss step to generating
views which the circular view uses. I may add it to the airport view as well.