Use selection filters instead of css selectors
I'm worried that the ids will not always encode the ids correctly. To deal with
this I build a selection using a filter. This is slower for sure than a css
selection. I may instead build a data structure to handle this stuff later. It
would be faster and not worry about encoding, but would use more memory.