Rename jmol component to jmolTools
We are going to have a pure jmol interface soon. In order to have both I'm
renaming the jmol component to jmolTools, because it uses jmolTools as an
interface to jmol. While this works great for many things, it can be a problem
when we want to have a whole structure colored. So I'm going to use pure jmol
for that.