First pass at a pure jmol interface
This is a pure jmol interface for the library. This does not use jmol tools to
laod things, but instead it will load a whole structure and let you work with
that. This is another way of working with the 3D structures and is probably
more helpful to coloring tasks and the like. However this can't do things like
show neighborhood and the like.