Large refactor of drawing and highlighting
Now nucleotides are highlighted in a consitent way across views. This has
promoted the addition of a highlights component. In addition, the way things
are drawn is changed. Now each view defines a callback that handles only adding
the data it must for nts, chains, interactions, etc. All the standard parts are
defined in the parent view class.