Basic attempt at adding helix labels
In the case of airport diagram we simply put the labels where we are asked to.
We can't correctly compute a location. In the case of a circular diagram we
place the labels at the first base in the helix. This is limited and not the
best thing to do long term. However, we are starting with it to have something
working.