Add highlighting to helix labels
Now mousing over a helix label can enable highlighting all letters in the
helix. This is done through the same mechanism as with all other mouseovers.
Right now the highlighting needs work as it is hard to read because letters
overlap. I also refactored how the mouseover and mouseout handlers are
generated.