First pass at using browserify
I'm trying to use a more modern and scalable approach to building this library.
For that end I've chosen browserify which lets me have modules and such. This
should allow easier testing and simplify the structuring of my code. Right now
it is still very similar to the previous structure, but I may fix that later.