Start of using es6
I'm moving to use es6 and babel in this project. I think it's cleaner and an
easier to use version of javascript. Plus it's the future and mostly supported
anyway. This requires changes to all of the code. Not everything has been
updated yet but I'm working through it. I'm also adding tests for things as I
go. I will be using mocha as it has support for using babel and it is a nice way
of writing tests.